+++ src/zrb/llm/task/llm_chat_task.py
@@ -24,6 +24,8 @@ from zrb.llm.prompt.tool_guidance import ToolGuidance
 from zrb.llm.summarizer import (
     create_summarizer_history_processor,
 )
+from zrb.llm.task._chat_builder_mixin import LLMChatBuilderMixin
+from zrb.llm.task._chat_runner_mixin import LLMChatRunnerMixin
 from zrb.llm.task.llm_task import LLMTask
 from zrb.llm.tool_call import (
     ArgumentFormatter,
@@ -79,7 +81,7 @@ def _parse_yolo_value(value: Any) -> "bool | frozenset[str]":
     return tools if tools else False
 
 
-class LLMChatTask(BaseTask):
+class LLMChatTask(LLMChatBuilderMixin, LLMChatRunnerMixin, BaseTask):
 
     def __init__(
         self,
@@ -352,191 +354,6 @@ class LLMChatTask(BaseTask):
         self._interactive = interactive
         self._show_ollama_models = show_ollama_models
         self._show_pydantic_ai_models = show_pydantic_ai_models
-
-    @property
-    def prompt_manager(self) -> PromptManager:
-        if self._prompt_manager is None:
-            raise ValueError(f"Task {self.name} doesn't have prompt_manager")
-        return self._prompt_manager
-
-    def set_ui(self, ui: "UIProtocol | list[UIProtocol] | None"):
-        """Set the UI protocol(s) for this task."""
-        self._uis = [] if ui is None else (ui if isinstance(ui, list) else [ui])
-
-    def append_ui(self, ui: "UIProtocol") -> None:
-        """Append a UI to the list of UIs."""
-        self._uis.append(ui)
-
-    def set_ui_factory(self, ui_factory: Callable[..., "UIProtocol"] | None):
-        """Set a factory function to instantiate the UI dynamically during execution."""
-        self._ui_factories = [] if ui_factory is None else [ui_factory]
-
-    def append_ui_factory(self, factory: Callable[..., "UIProtocol"]) -> None:
-        """Append a UI factory to the list of factories."""
-        self._ui_factories.append(factory)
-
-    def set_history_manager(self, history_manager: "AnyHistoryManager") -> None:
-        """Set the history manager for this task."""
-        self._history_manager = history_manager
-
-    @property
-    def custom_model_names(self) -> "StrListAttr | None":
-        return self._custom_model_names
-
-    @custom_model_names.setter
-    def custom_model_names(self, value: "StrListAttr | None"):
-        self._custom_model_names = value
-
-    @property
-    def model_getter(
-        self,
-    ) -> "Callable[[Model | str | None], Model | str | None] | None":
-        return self._model_getter
-
-    @model_getter.setter
-    def model_getter(
-        self, value: "Callable[[Model | str | None], Model | str | None] | None"
-    ):
-        self._model_getter = value
-
-    @property
-    def model_renderer(
-        self,
-    ) -> "Callable[[Model | str | None], Model | str | None] | None":
-        return self._model_renderer
-
-    @model_renderer.setter
-    def model_renderer(
-        self, value: "Callable[[Model | str | None], Model | str | None] | None"
-    ):
-        self._model_renderer = value
-
-    def set_approval_channel(self, channel: "ApprovalChannel | None"):
-        """Set the approval channel for tool confirmations."""
-        self._approval_channels = [] if channel is None else [channel]
-
-    def append_approval_channel(self, channel: "ApprovalChannel") -> None:
-        """Append an approval channel to the list."""
-        self._approval_channels.append(channel)
-
-    def add_toolset(self, *toolset: AbstractToolset):
-        self.append_toolset(*toolset)
-
-    def append_toolset(self, *toolset: AbstractToolset):
-        self._toolsets += list(toolset)
-
-    def add_toolset_factory(
-        self, *factory: Callable[[AnyContext], AbstractToolset[None]]
-    ):
-        self.append_toolset_factory(*factory)
-
-    def append_toolset_factory(
-        self, *factory: Callable[[AnyContext], AbstractToolset[None]]
-    ):
-        self._toolset_factories += list(factory)
-
-    def add_tool(self, *tool: Tool | ToolFuncEither):
-        self.append_tool(*tool)
-
-    def append_tool(self, *tool: Tool | ToolFuncEither):
-        self._tools += list(tool)
-
-    def add_tool_factory(self, *factory: Callable[[AnyContext], Tool | ToolFuncEither]):
-        self.append_tool_factory(*factory)
-
-    def append_tool_factory(
-        self, *factory: Callable[[AnyContext], Tool | ToolFuncEither]
-    ):
-        self._tool_factories += list(factory)
-
-    def add_tool_guidance_factory(
-        self,
-        *guidance_factory: Callable[[AnyContext], ToolGuidance],
-    ):
-        self.append_tool_guidance_factory(*guidance_factory)
-
-    def append_tool_guidance_factory(
-        self,
-        *guidance_factory: Callable[[AnyContext], ToolGuidance],
-    ):
-        """Register guidance for dynamically-named factory tools.
-
-        The factory is called when tools are resolved from factories. It should
-        return a single ToolGuidance object.
-        """
-        self._tool_guidance_factories += list(guidance_factory)
-
-    def add_tool_guidance(self, *guidance: ToolGuidance):
-        self.append_tool_guidance(*guidance)
-
-    def append_tool_guidance(self, *guidance: ToolGuidance):
-        """Add tool guidance entries to be applied when prompt_manager is available."""
-        self._pending_tool_guidance.extend(guidance)
-
-    def _apply_tool_guidance(self):
-        """Apply all pending tool guidance to the prompt manager."""
-        if self._prompt_manager is None:
-            return
-        for g in self._pending_tool_guidance:
-            self._prompt_manager.add_tool_guidance(
-                group=g.group_name,
-                name=g.tool_name,
-                use_when=g.when_to_use,
-                key_rule=g.key_rule,
-            )
-        self._pending_tool_guidance.clear()
-
-    def add_hook_factory(self, *factory: Callable[[HookManager], None]):
-        self.append_hook_factory(*factory)
-
-    def append_hook_factory(self, *factory: Callable[[HookManager], None]):
-        self._hook_factories += list(factory)
-
-    def add_history_processor(self, *processor: HistoryProcessor):
-        self.append_history_processor(*processor)
-
-    def append_history_processor(self, *processor: HistoryProcessor):
-        self._history_processors += list(processor)
-
-    def add_response_handler(self, *handler: ResponseHandler):
-        self.prepend_response_handler(*handler)
-
-    def prepend_response_handler(self, *handler: ResponseHandler):
-        self._response_handlers = list(handler) + self._response_handlers
-
-    def add_tool_policy(self, *policy: ToolPolicy):
-        self.prepend_tool_policy(*policy)
-
-    def prepend_tool_policy(self, *policy: ToolPolicy):
-        self._tool_policies = list(policy) + self._tool_policies
-
-    def add_argument_formatter(self, *formatter: ArgumentFormatter):
-        self.prepend_argument_formatter(*formatter)
-
-    def prepend_argument_formatter(self, *formatter: ArgumentFormatter):
-        self._argument_formatters = list(formatter) + self._argument_formatters
-
-    def add_trigger(self, *trigger: Callable[[], AsyncIterable[Any]]):
-        self.append_trigger(*trigger)
-
-    def append_trigger(self, *trigger: Callable[[], AsyncIterable[Any]]):
-        self._triggers += trigger
-
-    def add_custom_command(
-        self,
-        *custom_command: (
-            AnyCustomCommand | Callable[[], AnyCustomCommand | list[AnyCustomCommand]]
-        ),
-    ):
-        self.append_custom_command(*custom_command)
-
-    def append_custom_command(
-        self,
-        *custom_command: (
-            AnyCustomCommand | Callable[[], AnyCustomCommand | list[AnyCustomCommand]]
-        ),
-    ):
-        self._custom_commands += list(custom_command)
 
     async def _exec_action(self, ctx: AnyContext) -> Any:
         # 1. Resolve inputs/attributes
@@ -844,257 +661,6 @@ class LLMChatTask(BaseTask):
             model_renderer=self._model_renderer,
             summarize_command=summarize_commands,
         )
-
-    async def _run_non_interactive_session(
-        self,
-        ctx: AnyContext,
-        llm_task_core: LLMTask,
-        initial_message: Any,
-        initial_conversation_name: str,
-        initial_yolo: bool | frozenset[str],
-        initial_attachments: list[UserContent],
-    ) -> Any:
-        # AsyncExitStack is handled by LLMTask._exec_action
-        session_input = {
-            "message": initial_message,
-            "session": initial_conversation_name,
-            "yolo": bool(initial_yolo),  # inner task uses dynamic_yolo; just pass bool
-            "attachments": initial_attachments,
-            "model": self._get_model(ctx),
-        }
-        shared_ctx = SharedContext(
-            input=session_input,
-            print_fn=ctx.shared_print,  # Use current task's print function
-        )
-        session = Session(shared_ctx)
-        result = await llm_task_core.async_run(session)
-        # Store conversation name in xcom for CLI to print at the end
-        ctx.xcom["__conversation_name__"] = initial_conversation_name
-        return result
-
-    async def _run_interactive_session(
-        self,
-        ctx: AnyContext,
-        llm_task_core: LLMTask,
-        history_manager: AnyHistoryManager,
-        ui_commands: dict[str, list[str]],
-        initial_message: Any,
-        initial_conversation_name: str,
-        initial_yolo: bool | frozenset[str],
-        initial_attachments: list[UserContent],
-        enable_rewind: bool = False,
-        snapshot_dir: str = "",
-    ) -> Any:
-        from zrb.llm.ui.base_ui import BaseUI
-
-        # Note: AsyncExitStack is handled by LLMTask._exec_action
-        # 1. Resolve UIs from factories
-        resolved_uis: list["UIProtocol"] = list(self._uis)
-        for factory in self._ui_factories:
-            factory_ui = factory(
-                ctx=ctx,
-                llm_task=llm_task_core,
-                history_manager=history_manager,
-                ui_commands=ui_commands,
-                initial_message=initial_message,
-                initial_conversation_name=initial_conversation_name,
-                initial_yolo=initial_yolo,
-                initial_attachments=initial_attachments,
-            )
-            if isinstance(factory_ui, list):
-                resolved_uis.extend(factory_ui)
-            else:
-                resolved_uis.append(factory_ui)
-
-        # 2. Resolve UI attributes for default UI
-        resolved_custom_model_names = get_attr(ctx, self._custom_model_names, []) or []
-        if not isinstance(resolved_custom_model_names, list):
-            resolved_custom_model_names = []
-        ui_greeting = get_str_attr(ctx, self._ui_greeting, "", self._render_ui_greeting)
-        ui_assistant_name = get_str_attr(
-            ctx, self._ui_assistant_name, "", self._render_ui_assistant_name
-        )
-        ui_jargon = get_str_attr(ctx, self._ui_jargon, "", self._render_ui_jargon)
-        ascii_art = get_str_attr(
-            ctx, self._ui_ascii_art_name, "", self._render_ui_ascii_art_name
-        )
-        # Resolve show_ollama_models and show_pydantic_ai_models with CFG fallback
-        effective_show_ollama_models = (
-            CFG.LLM_SHOW_OLLAMA_MODELS
-            if self._show_ollama_models is None
-            else self._show_ollama_models
-        )
-        effective_show_pydantic_ai_models = (
-            CFG.LLM_SHOW_PYDANTIC_AI_MODELS
-            if self._show_pydantic_ai_models is None
-            else self._show_pydantic_ai_models
-        )
-
-        # Resolve custom commands
-        resolved_custom_commands: list[AnyCustomCommand] = []
-        for cmd in self._custom_commands:
-            if callable(cmd):
-                res = cmd()
-                if isinstance(res, list):
-                    resolved_custom_commands.extend(res)
-                else:
-                    resolved_custom_commands.append(res)
-            else:
-                resolved_custom_commands.append(cmd)
-
-        # 3. Determine the UI to use
-        from zrb.llm.app.lexer import CLIStyleLexer
-        from zrb.llm.ui.default_ui import UI
-        from zrb.llm.ui.multi_ui import MultiUI
-
-        ui: "UIProtocol | None" = None
-
-        if resolved_uis and not self._include_default_ui:
-            # Factory UIs only — skip the default CLI UI entirely
-            if len(resolved_uis) == 1:
-                ui = resolved_uis[0]
-            else:
-                ui = MultiUI(resolved_uis)
-                if len(self._approval_channels) == 1:
-                    ui.set_approval_channel(self._approval_channels[0])
-                elif len(self._approval_channels) > 1:
-                    from zrb.llm.approval import MultiplexApprovalChannel
-
-                    ui.set_approval_channel(
-                        MultiplexApprovalChannel(self._approval_channels)
-                    )
-        elif resolved_uis:
-            # We have factory UIs - create default UI and combine them
-            default_ui = UI(
-                ctx=ctx,
-                yolo_xcom_key=self._yolo_xcom_key,
-                greeting=ui_greeting,
-                assistant_name=ui_assistant_name,
-                ascii_art=ascii_art,
-                jargon=ui_jargon,
-                output_lexer=CLIStyleLexer(),
-                llm_task=llm_task_core,
-                history_manager=history_manager,
-                initial_message=initial_message,
-                initial_attachments=initial_attachments,
-                conversation_session_name=initial_conversation_name,
-                is_yolo=initial_yolo,
-                triggers=self._triggers,
-                response_handlers=self._response_handlers,
-                tool_policies=self._tool_policies,
-                argument_formatters=self._argument_formatters,
-                markdown_theme=self._markdown_theme,
-                summarize_commands=ui_commands["summarize"],
-                attach_commands=ui_commands["attach"],
-                exit_commands=ui_commands["exit"],
-                info_commands=ui_commands["info"],
-                save_commands=ui_commands["save"],
-                load_commands=ui_commands["load"],
-                rewind_commands=ui_commands["rewind"],
-                yolo_toggle_commands=ui_commands["yolo_toggle"],
-                set_model_commands=ui_commands["set_model"],
-                redirect_output_commands=ui_commands["redirect_output"],
-                exec_commands=ui_commands["exec"],
-                btw_commands=ui_commands["btw"],
-                custom_commands=resolved_custom_commands,
-                model=self._get_model(ctx),
-                custom_model_names=resolved_custom_model_names,
-                show_ollama_models=effective_show_ollama_models,
-                show_pydantic_ai_models=effective_show_pydantic_ai_models,
-                enable_rewind=enable_rewind,
-                snapshot_dir=snapshot_dir,
-            )
-            # Add default UI first, then factory UIs
-            all_uis = [default_ui] + resolved_uis
-            if len(all_uis) == 1:
-                ui = default_ui
-            else:
-                ui = MultiUI(all_uis)
-                # Set up approval channel for MultiUI
-                if len(self._approval_channels) == 1:
-                    ui.set_approval_channel(self._approval_channels[0])
-                elif len(self._approval_channels) > 1:
-                    from zrb.llm.approval import MultiplexApprovalChannel
-
-                    ui.set_approval_channel(
-                        MultiplexApprovalChannel(self._approval_channels)
-                    )
-                # Set tool call handler so CLI mode has same formatters as standalone CLI
-                ui.set_tool_call_handler(default_ui.tool_call_handler)
-        else:
-            # No factory UIs, use default UI
-            ui = UI(
-                ctx=ctx,
-                yolo_xcom_key=self._yolo_xcom_key,
-                greeting=ui_greeting,
-                assistant_name=ui_assistant_name,
-                ascii_art=ascii_art,
-                jargon=ui_jargon,
-                output_lexer=CLIStyleLexer(),
-                llm_task=llm_task_core,
-                history_manager=history_manager,
-                initial_message=initial_message,
-                initial_attachments=initial_attachments,
-                conversation_session_name=initial_conversation_name,
-                is_yolo=initial_yolo,
-                triggers=self._triggers,
-                response_handlers=self._response_handlers,
-                tool_policies=self._tool_policies,
-                argument_formatters=self._argument_formatters,
-                markdown_theme=self._markdown_theme,
-                summarize_commands=ui_commands["summarize"],
-                attach_commands=ui_commands["attach"],
-                exit_commands=ui_commands["exit"],
-                info_commands=ui_commands["info"],
-                save_commands=ui_commands["save"],
-                load_commands=ui_commands["load"],
-                rewind_commands=ui_commands["rewind"],
-                yolo_toggle_commands=ui_commands["yolo_toggle"],
-                set_model_commands=ui_commands["set_model"],
-                redirect_output_commands=ui_commands["redirect_output"],
-                exec_commands=ui_commands["exec"],
-                btw_commands=ui_commands["btw"],
-                custom_commands=resolved_custom_commands,
-                model=self._get_model(ctx),
-                custom_model_names=resolved_custom_model_names,
-                show_ollama_models=effective_show_ollama_models,
-                show_pydantic_ai_models=effective_show_pydantic_ai_models,
-                enable_rewind=enable_rewind,
-                snapshot_dir=snapshot_dir,
-            )
-
-        # 4. Load and display session history if session name is provided
-        if initial_conversation_name:
-            try:
-                from zrb.llm.util.history_formatter import format_history_as_text
-
-                history = history_manager.load(initial_conversation_name)
-                if history:
-                    history_text = format_history_as_text(history)
-                    ui.append_to_output(history_text)
-            except FileNotFoundError:
-                # New session - no history file exists yet
-                pass
-            except Exception as e:
-                CFG.LOGGER.warning(
-                    f"Failed to load history for session {initial_conversation_name}: {e}"
-                )
-
-        # 5. Run the UI
-        if ui is None:
-            raise ValueError("No UI available")
-        if isinstance(ui, BaseUI) or hasattr(ui, "run_async"):
-            await ui.run_async()
-        else:
-            raise ValueError(f"UI {type(ui)} does not implement run_async")
-        last_output = getattr(ui, "last_output", "")
-        final_conversation_name = self._get_ui_conversation_name(
-            ui, initial_conversation_name
-        )
-        # Print session name at the very end (after result is returned to caller)
-        # Store it to be printed by the finally block in cli.py
-        ctx.xcom["__conversation_name__"] = final_conversation_name
-        return last_output
 
     def _print_conversation_name(self, ctx: AnyContext, conversation_name: str):
         stylized_label = stylize_faint("Session")
